Add win_opts.row and win_opts.col
The idea comes from the fork inferst/screenkey.nvim,
where the position of the window is hard-coded. Some streamers have
their camera in the bottom left corner of the screen, so you cannot
see the plugin in action.

I also added a note into README, describing how to move the window
to the top left corner.