feat(api): provide `redraw` function
First of all it's much straight-forward
to call `redraw` then to call `toggle` twice.
`redraw` fixes some issues, for example if the
Screenkey window needs to be closed and reopened
when some filetype from `display_infront` is opened,
then the text inside Screenkey window will be cleared.
With redraw this is no longer the case. Also,
this function could be useful for other use cases.